DEF-2145. It takes a lot of time to open a project with a bunch of dependencies
* resolveLibUrls checks SHA1 on code 200 as github replies this even
with a matching etag (as oppsosed to 304). 40% of the problem.
* linkLibraries was using IResource.createLink with an URI, which from
what I can tell tries to resolve the URI, taking a lot of time (the
remaining 60%). This is using an Path instead, since we’re linking
local resources (libs).

+++ com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/Project.java
@@ -743,14 +743,26 @@ run:
                 if (code == 304) {
                     // Reusing cached library
                 } else {
-                    input = new BufferedInputStream(connection.getInputStream());
-                    FileUtils.copyInputStreamToFile(input, f);
-                    try {
-                        ZipFile zip = new ZipFile(f);
-                        zip.close();
-                    } catch (ZipException e) {
-                        f.delete();
-                        throw new LibraryException(String.format("The file obtained from %s is not a valid zip file", url.toString()), e);
+                    boolean serverSha1Match = false;
+                    if(code == 200) {
+                        // GitHub uses eTags and we can check we have the up to date version by comparing SHA1 and server eTag if we get a 200 OK response
+                        String serverETag = connection.getHeaderField("ETag").replace("\"", "");
+                        if((sha1 != null) && (sha1.equals(serverETag))) {
+                            // Reusing cached library
+                           serverSha1Match = true;
+                        }
+                    }
+                    if(!serverSha1Match) {
+                        input = new BufferedInputStream(connection.getInputStream());
+                        FileUtils.copyInputStreamToFile(input, f);
+
+                        try {
+                            ZipFile zip = new ZipFile(f);
+                            zip.close();
+                        } catch (ZipException e) {
+                            f.delete();
+                            throw new LibraryException(String.format("The file obtained from %s is not a valid zip file", url.toString()), e);
+                        }
                     }
                 }
                 connection.disconnect();
@@ -759,7 +771,9 @@ run:
             } catch (FileNotFoundException e) {
                 throw new LibraryException(String.format("The URL %s points to a resource which doesn't exist", url.toString()), e);
             } finally {
-                IOUtils.closeQuietly(input);
+                if(input != null) {
+                    IOUtils.closeQuietly(input);
+                }
                 subProgress.worked(1);
             }
         }
